requirements.txt
dj-rest-auth==2.2.5
Django==3.2.16
django-allauth==0.51.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
requests==2.27.1
requests-oauthlib==1.3.1